+++ modules/client/test/integration/addresslistVoting-client/methods.test.ts
@@ -22,6 +22,7 @@ import {
 } from "../../../src";
 import {
   ADDRESS_ONE,
+  ADDRESS_TWO,
   contextParamsLocalChain,
   SUBGRAPH_ACTIONS,
   SUBGRAPH_PLUGIN_INSTALLATION,
@@ -42,6 +43,7 @@ import {
 import { buildAddressListVotingDAO } from "../../helpers/build-daos";
 import { JsonRpcProvider } from "@ethersproject/providers";
 import {
+  QueryAddresslistVotingIsMember,
   QueryAddresslistVotingMembers,
   QueryAddresslistVotingProposal,
   QueryAddresslistVotingProposals,
@@ -53,13 +55,13 @@ import {
 } from "../../../src/addresslistVoting/internal/types";
 import {
   Context,
+  getExtendedProposalId,
   InvalidAddressOrEnsError,
   PrepareInstallationStep,
   PrepareUpdateStep,
   ProposalMetadata,
   ProposalStatus,
   SortDirection,
-  getExtendedProposalId,
 } from "@aragon/sdk-client-common";
 import { PluginRepo__factory } from "@aragon/osx-ethers";
 import { createAddresslistVotingPluginBuild } from "../../helpers/create-plugin-build";
@@ -681,6 +683,86 @@ describe("Client Address List", () => {
             );
         }
       }
+    });
+  });
+
+  describe("isMember", () => {
+    it("Should check if an address is a member of a DAO and return true", async () => {
+      const ctx = new Context(contextParamsLocalChain);
+      const client = new AddresslistVotingClient(ctx);
+
+      const mockedClient = mockedGraphqlRequest.getMockedInstance(
+        client.graphql.getClient(),
+      );
+      mockedClient.request.mockResolvedValueOnce({
+        addresslistVotingVoter: {
+          id: `${ADDRESS_ONE}_${ADDRESS_TWO}`,
+        },
+      });
+      const isMember = await client.methods.isMember({
+        pluginAddress: ADDRESS_ONE,
+        address: ADDRESS_TWO,
+      });
+      expect(isMember).toBe(true);
+      expect(mockedClient.request).toHaveBeenLastCalledWith(
+        QueryAddresslistVotingIsMember,
+        {
+          id: `${ADDRESS_ONE}_${ADDRESS_TWO}`,
+          blockHeight: null,
+        },
+      );
+    });
+    it("Should check if an address is a member in a specified block number of a DAO and return true", async () => {
+      const ctx = new Context(contextParamsLocalChain);
+      const client = new AddresslistVotingClient(ctx);
+
+      const mockedClient = mockedGraphqlRequest.getMockedInstance(
+        client.graphql.getClient(),
+      );
+      mockedClient.request.mockResolvedValueOnce({
+        addresslistVotingVoter: {
+          id: `${ADDRESS_ONE}_${ADDRESS_TWO}`,
+        },
+      });
+      const blockNumber = 123456;
+      const isMember = await client.methods.isMember({
+        pluginAddress: ADDRESS_ONE,
+        address: ADDRESS_TWO,
+        blockNumber,
+      });
+      expect(isMember).toBe(true);
+      expect(mockedClient.request).toHaveBeenLastCalledWith(
+        QueryAddresslistVotingIsMember,
+        {
+          id: `${ADDRESS_ONE}_${ADDRESS_TWO}`,
+          blockHeight: {
+            number: blockNumber,
+          },
+        },
+      );
+    });
+    it("Should check if an address is a member of a DAO and return false", async () => {
+      const ctx = new Context(contextParamsLocalChain);
+      const client = new AddresslistVotingClient(ctx);
+
+      const mockedClient = mockedGraphqlRequest.getMockedInstance(
+        client.graphql.getClient(),
+      );
+      mockedClient.request.mockResolvedValueOnce({
+        addresslistVotingVoter: null,
+      });
+      const isMember = await client.methods.isMember({
+        pluginAddress: ADDRESS_ONE,
+        address: ADDRESS_TWO,
+      });
+      expect(isMember).toBe(false);
+      expect(mockedClient.request).toHaveBeenLastCalledWith(
+        QueryAddresslistVotingIsMember,
+        {
+          id: `${ADDRESS_ONE}_${ADDRESS_TWO}`,
+          blockHeight: null,
+        },
+      );
     });
   });